requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
pandas>=2.0.0
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
matplotlib>=3.7.0
